refactor(puretransformer): replaced built-in pureTransformer code with the library js-image-lib
Moved code from pureTransformer to its own npm package
