+++ package.json
@@ -47,17 +47,10 @@
     "release": "dotenv release-it --"
   },
   "dependencies": {
-    "@pdf-lib/upng": "^1.0.1",
-    "@wokwi/bmp-ts": "^3.0.0",
-    "buffer": "^6.0.3",
     "clsx": "^1.1.1",
-    "get-rgba-palette": "^2.0.1",
     "is-svg": "^4.3.2",
-    "jpeg-js": "^0.4.3",
-    "omggif": "^1.0.10",
-    "pako": "^2.0.4",
-    "query-string": "^7.1.0",
-    "utif": "^3.1.0"
+    "js-image-lib": "^0.1.3",
+    "query-string": "^7.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.0",
@@ -80,11 +73,7 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.0.3",
-    "@types/omggif": "^1.0.2",
-    "@types/pngjs": "^6.0.1",
-    "@types/qs": "^6.9.7",
     "@types/react": "^17.0.37",
-    "@types/utif": "^3.0.1",
     "@typescript-eslint/eslint-plugin": "^5.5.0",
     "@typescript-eslint/parser": "^5.5.0",
     "autoprefixer": "^10.4.0",
