+++ src/components/Image/index.tsx
@@ -1,2 +1,1 @@
-export { Image } from "./Image";
-export type { ImageProps } from "./Image";
+export * from "./Image";
