--- src/hooks/responsiveImage/index.ts
@@ -1,2 +1,0 @@
-export { useResponsiveImage } from "./responsiveImage";
-export type { ResponsiveHookResult } from "./responsiveImage";
