--- src/server/caches/diskCache/index.ts
@@ -1,2 +1,0 @@
-export { DiskCache } from "./diskCache";
-export type { DiskCacheConfig } from "./diskCache";
