--- src/server/caches/memoryCache/index.ts
@@ -1,2 +1,0 @@
-export { MemoryCache } from "./memoryCache";
-export type { MemoryCacheConfig } from "./memoryCache";
