--- src/server/loaders/imageLoader/index.ts
@@ -1,1 +1,0 @@
-export { imageLoader } from "./imageLoader";
