+++ src/server/pure.ts
@@ -3,4 +3,4 @@ export * from "./transformers/pureTransformer";
 export * from "./resolvers/fetchResolver";
 export * from "./resolvers/kvResolver";
 export * from "./loaders";
-export * from "../utils/cache";
+export * from "../utils";
