--- src/server/resolvers/fetchResolver/index.ts
@@ -1,1 +1,0 @@
-export { fetchResolver } from "./fetchResolver";
