--- src/server/resolvers/fsResolver/index.ts
@@ -1,1 +1,0 @@
-export { fsResolver } from "./fsResolver";
