--- src/server/resolvers/kvResolver/index.ts
@@ -1,1 +1,0 @@
-export { kvResolver } from "./kvResolver";
