+++ src/server/transformers/pureTransformer.ts
@@ -1,0 +1,25 @@
+import { imageTransformer } from "js-image-lib";
+import { MimeType } from "../../types/file";
+import { Transformer } from "../../types/transformer";
+
+const supportedInputs = new Set([
+  MimeType.JPEG,
+  MimeType.PNG,
+  MimeType.GIF,
+  MimeType.BMP,
+  MimeType.TIFF,
+]);
+
+const supportedOutputs = new Set([
+  MimeType.JPEG,
+  MimeType.PNG,
+  MimeType.GIF,
+  MimeType.BMP,
+]);
+
+export const pureTransformer: Transformer = {
+  name: "pureTransformer",
+  supportedInputs,
+  supportedOutputs,
+  transform: async (src, options) => imageTransformer(src, options),
+};
