--- src/server/transformers/pureTransformer/index.ts
@@ -1,1 +1,0 @@
-export { pureTransformer } from "./pureTransformer";
