+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./cache";
+export * from "./fileType";
